requirements.txt
# Polymarket Insider Alert Bot - Dependencies
httpx>=0.25.0
python-telegram-bot>=20.7
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
structlog>=23.2.0
